* Add menu items to the form. * Move PropertyControl.xaml to Controls folder
