+++ Factory Ide/Factory Ide/Controls/PropertyControl.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="Factory_Ide.PropertyControl"
+﻿<UserControl x:Class="Factory_Ide.Controls.PropertyControl"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
